+++ lib/src/utils/app_theme.dart
@@ -1,5 +1,4 @@
-import 'package:flutter/material.dart';
-import 'package:itq_utils/itq_utils.dart';
+import 'dart:ui';
 
 class AppTheme {
   AppTheme._();
@@ -22,7 +21,7 @@ class AppTheme {
   static const Color spacer = Color(0xFFF2F2F2);
   static const String fontName = 'Roboto';
 
-  static const TextStyle display1 = TextStyle(
+  static TextStyle display1 = TextStyle(
     fontFamily: fontName,
     fontWeight: FontWeight.bold,
     fontSize: 36,
@@ -31,7 +30,7 @@ class AppTheme {
     color: darkerText,
   );
 
-  static const TextStyle headline = TextStyle(
+  static TextStyle headline = TextStyle(
     fontFamily: fontName,
     fontWeight: FontWeight.bold,
     fontSize: 24,
@@ -39,7 +38,7 @@ class AppTheme {
     color: darkerText,
   );
 
-  static const TextStyle title = TextStyle(
+  static TextStyle title = TextStyle(
     fontFamily: fontName,
     fontWeight: FontWeight.bold,
     fontSize: 16,
@@ -47,7 +46,7 @@ class AppTheme {
     color: darkerText,
   );
 
-  static const TextStyle subtitle = TextStyle(
+  static TextStyle subtitle = TextStyle(
     fontFamily: fontName,
     fontWeight: FontWeight.w400,
     fontSize: 14,
@@ -55,7 +54,7 @@ class AppTheme {
     color: darkText,
   );
 
-  static const TextStyle body2 = TextStyle(
+  static TextStyle body2 = TextStyle(
     fontFamily: fontName,
     fontWeight: FontWeight.w400,
     fontSize: 14,
@@ -63,7 +62,7 @@ class AppTheme {
     color: darkText,
   );
 
-  static const TextStyle body1 = TextStyle(
+  static TextStyle body1 = TextStyle(
     fontFamily: fontName,
     fontWeight: FontWeight.w400,
     fontSize: 16,
@@ -71,7 +70,7 @@ class AppTheme {
     color: darkText,
   );
 
-  static const TextStyle caption = TextStyle(
+  static TextStyle caption = TextStyle(
     fontFamily: fontName,
     fontWeight: FontWeight.w400,
     fontSize: 12,