+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-## [1.0.1] -  Planned Updates
+## [1.0.2] -  Planned Updates
 ### Improved
+- Enhanced image uploader functionality and material widgets for **FluentUploadImage**.
+- Enhanced image uploader functionality and material widgets for **FluentDropdownSearch**.
+
+## [1.0.1] -  Planned Updates
+### Improved``
 - Enhanced image uploader functionality for **FluentUploadImage**.
 
 ## [1.0.0] - Initial Release